(svn r1288) -Codechange: changed _map2 to an uint16. It is still saved and loaded as an uint8 till the savegame version is bumped to version 5. Then it works automaticly as a fully uint16. So _stations[] can not be increased till after the bump!!

+++ industry_cmd.c
@@ -837,7 +837,7 @@ void DeleteIndustry(Industry *i)
 
 	BEGIN_TILE_LOOP(tile_cur,	i->width, i->height, i->xy);
 		if (IS_TILETYPE(tile_cur, MP_INDUSTRY)) {
-			if (_map2[tile_cur] == (byte)index) {
+			if (_map2[tile_cur] == (uint16)index) {
 				DoClearSquare(tile_cur);
 			}
 		} else if (IS_TILETYPE(tile_cur, MP_STATION) && _map5[tile_cur] == 0x4B) {
