+++ map.h
@@ -10,13 +10,13 @@
 #define TILE_X_MAX (TILES_X - 1)
 #define TILE_Y_MAX (TILES_Y - 1)
 
-extern byte _map_type_and_height[];
-extern byte _map5[];
-extern byte _map3_lo[];
-extern byte _map3_hi[];
-extern byte _map_owner[];
-extern byte _map2[];
-extern byte _map_extra_bits[];
+extern byte   _map_type_and_height[];
+extern byte   _map5[];
+extern byte   _map3_lo[];
+extern byte   _map3_hi[];
+extern byte   _map_owner[];
+extern uint16 _map2[];
+extern byte   _map_extra_bits[];
 
 // binary logarithm of the map size, try to avoid using this one
 static inline uint MapLogX(void)  { extern uint _map_log_x; return _map_log_x; }
